rename for splitting entity Child and Parent
