+++ components/ParentPage/SideBar/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import styled from 'styled-components';
-import ParentCard from '../../UsersCard/ParentCard';
-import ChildCard from '../../UsersCard/ChildCard';
+import ParentCard from './UsersCard/ParentCard';
+import ChildCard from './UsersCard/ChildCard';
 import { users } from '../../../const';
 import AvatarLily from '../../../src/icons/Lily.svg';
 import AvatarMary from '../../../src/icons/Mary.svg';
